services/settlement-service/requirements.txt
# Settlement Service — Python dependencies
# All versions pinned for reproducibility and security (OWASP A06).
# Update with: pip-compile requirements.in

# ── Web framework ──────────────────────────────────────────────
fastapi==0.110.2
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Database ───────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# ── Kafka ──────────────────────────────────────────────────────
aiokafka==0.10.0

# ── HTTP client (fraud service calls) ─────────────────────────
httpx==0.27.0

# ── Observability ──────────────────────────────────────────────
prometheus-fastapi-instrumentator==7.0.0

# OpenTelemetry — distributed tracing
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-httpx==0.45b0
opentelemetry-instrumentation-sqlalchemy==0.45b0

# ── Utilities ──────────────────────────────────────────────────
python-dotenv==1.0.1
python-multipart==0.0.9

# ── Testing ────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
jsonschema==4.22.0     # Kafka contract tests
httpx==0.27.0          # async test client
aiosqlite==0.20.0      # SQLite async driver (in-memory test DB)
factory-boy==3.3.0

# ── Linting / Security ─────────────────────────────────────────
ruff==0.4.1
black==24.4.2
mypy==1.9.0
bandit[toml]==1.7.8
